Move more post-processing structures to here.
